+++ tasks/generate-fixture-html.js
@@ -1,16 +1,15 @@
 const { globSync } = require("glob");
 const fs = require("fs");
-const path = require("path");
-const Diff = require("diff");
 const { renderNunjucks } = require("./lib/nunjucks");
 
-const componentsDirectory = "src/nationalarchives/components/";
-const componentFixturesFile = "/fixtures.json";
 const fixturesOutputDirectory = "fixtures-html";
 
 if (!fs.existsSync(fixturesOutputDirectory)) {
   fs.mkdirSync(fixturesOutputDirectory);
 }
+
+const componentsDirectory = "src/nationalarchives/components/";
+const componentFixturesFile = "/fixtures.json";
 
 const components = globSync(
   `${componentsDirectory}*${componentFixturesFile}`,
@@ -32,7 +31,7 @@ components.forEach((component) => {
       params: fixture.options,
     });
     fs.writeFile(
-      `${fixturesOutputDirectory}/${component}-${fixture.name
+      `${fixturesOutputDirectory}/component-${component}-${fixture.name
         .replace(/[^0-9a-z]/gi, "-")
         .toLowerCase()}.html`,
       result,
@@ -44,3 +43,27 @@ components.forEach((component) => {
     );
   });
 });
+
+const templatesDirectory = "../src/nationalarchives/templates/";
+
+require(`${templatesDirectory}fixtures.json`)
+  .fixtures.filter((fixture) => fixture.template !== "email.njk")
+  .forEach((fixture) => {
+    const templateNunjucks = fixture.template
+      ? require(`${templatesDirectory}${fixture.template}`)
+      : fixture.string;
+    const result = renderNunjucks(templateNunjucks, {
+      ...fixture.options,
+    });
+    fs.writeFile(
+      `${fixturesOutputDirectory}/template-${fixture.name
+        .replace(/[^0-9a-z]/gi, "-")
+        .toLowerCase()}.html`,
+      result,
+      (err) => {
+        if (err) {
+          return console.log(err);
+        }
+      },
+    );
+  });